S3  works with basic; audio processing broke
on S3, file saved under UserID/FileID

+++ server/controllers/audios.js
@@ -3,10 +3,10 @@ const _ = require('lodash');
 const audiosRouter = require('express').Router();
 const User = require('../models/user');
 const Audio = require('../models/audio');
-const { getLoggedInUser, authorizeRequest } = require('../utils/authHelper');
+const { getLoggedInUser, tryAuthorizeRequest } = require('../utils/authHelper');
 const { getBufferFileFromId } = require('../utils/bufferHelper');
 const {
-  S3Delete, S3Insert, S3RetrieveAll, S3RetrieveItem,
+  S3Delete, S3Insert, S3Update,
 } = require('../utils/S3Storage');
 
 const upload = multer();
@@ -19,50 +19,59 @@ audiosRouter.get('/', async (request, response) => {
 });
 
 audiosRouter.get('/:id', async (request, response, next) => {
-  const user = await authorizeRequest(request, response);
-  await getBufferFileFromId(request, response, user.username);
+  if (await tryAuthorizeRequest(request, response)) {
+    await getBufferFileFromId(request, response);
+  }
 });
 
 audiosRouter.put('/:id', upload.single('test'), async (request, response, next) => {
-  await authorizeRequest(request, response);
-  const audio = await Audio.findById(request.params.id);
-  // S3Update()
-  await Audio.findByIdAndUpdate(request.params.id, { ...audio, date: new Date() }, { new: true });
-  response.status(200).end();
+  if (await tryAuthorizeRequest(request, response)) {
+    const audio = await Audio.findById(request.params.id);
+    await Audio.findByIdAndUpdate(request.params.id, { ...audio, date: new Date() }, { new: true });
+    try {
+      await S3Update(request.user.userID, request.params.id, request.file.buffer);
+    } catch (err) {
+      return response.status(400).send({ error: 'file override failed' });
+    }
+    response.status(200).end();
+  }
 });
 
 audiosRouter.delete('/:id', async (request, response, next) => {
-  await authorizeRequest(request, response);
-  
-  // Delete reference to audio from user
-  const audio = await Audio.findById(request.params.id);
-  const userId = audio.user;
-  const user = await User.findById(userId);
-  _.remove(user.audios, (audioID) => audioID === audio.id);
-  // await user.save(user);
-  user.save(user);
-  await S3Delete(user.username, request.params.id);
-  // Delete audio
-  await Audio.findByIdAndRemove(audio.id);
-  response.status(204).end();
+  if (await tryAuthorizeRequest(request, response)) {
+    // Delete reference to audio from user
+    const audio = await Audio.findById(request.params.id);
+    const userId = audio.user;
+    const user = await User.findById(userId);
+    _.remove(user.audios, (audioID) => audioID === audio.id);
+    await user.save(user);
+    try {
+      await S3Delete(userId, request.params.id);
+    } catch (err) {
+      return response.status(400).send({ error: 'file deletion failed' });
+    }
+    await Audio.findByIdAndRemove(audio.id);
+    response.status(204).end();
+  }
 });
 
 audiosRouter.post('/', upload.single('test'), async (request, response, next) => {
   const { name } = request.body;
   const { file } = request;
   const user = await getLoggedInUser(request, response);
-  
   const audio = new Audio({
     name,
-    content: file.buffer,
     user: user._id,
     date: new Date(),
   });
   const savedAudio = await audio.save();
+  try {
+    await S3Insert(request.user.userID, savedAudio._id, file.buffer);
+  } catch (err) {
+    return response.status(400).send({ error: 'file upload failed' });
+  }
   user.audios = user.audios.concat(savedAudio._id);
-  await S3Insert(savedAudio._id,file.buffer)
   await user.save();
-  await S3Insert(savedAudio._id, file.buffer);
   const res = { id: savedAudio._id };
   response.json(res);
 });
